Bug fixes: No students in database caused a crash with recommendations, and the planner crashed after deleting a semester.
Added a check to the recommendations view to only display the add recommendations button if there were students.

Added the destroy dependency to the semester model to delete all sections and applications when deleted.

Added a check to the custom drop down used only for the planner to only display when there are semesters.